requirements.txt
PySide6>=6.5.0
PySide6-Fluent-Widgets>=1.7.0
pypdf>=3.17.0
PyMuPDF>=1.23.0
reportlab>=4.0.0
python-docx>=1.0.0
odfpy>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0
loguru>=0.7.0